@@ -35,96 +33,54 @@
     "gaode"
   ],
   "author": "李梦龙",
-  "license": "ISC",
+  "license": "MIT",
   "bugs": {
     "url": "https://github.com/em-fe/em-map/issues"
   },
   "homepage": "https://github.com/em-fe/em-map#readme",
   "devDependencies": {
-    "autoprefixer": "^6.4.0",
-    "babel-core": "^6.17.0",
-    "babel-eslint": "^7.1.1",
-    "babel-helper-vue-jsx-merge-props": "^2.0.2",
-    "babel-loader": "^6.2.5",
-    "babel-plugin-syntax-jsx": "^6.8.0",
-    "babel-plugin-transform-runtime": "^6.0.0",
-    "babel-plugin-transform-vue-jsx": "^3.1.0",
-    "babel-preset-es2015": "^6.16.0",
-    "babel-preset-stage-0": "^6.16.0",
-    "babel-preset-stage-2": "^6.0.0",
-    "babel-register": "^6.0.0",
-    "babel-runtime": "^6.11.6",
-    "chai": "^3.5.0",
-    "codemirror": "^5.25.2",
-    "connect-history-api-fallback": "^1.1.0",
-    "copy-webpack-plugin": "^4.1.1",
-    "cross-env": "^3.1.3",
-    "cross-spawn": "^4.0.2",
-    "css-loader": "^0.28.0",
-    "eslint": "^3.7.1",
-    "eslint-config-elemefe": "^0.1.1",
-    "eslint-config-standard": "^6.1.0",
-    "eslint-friendly-formatter": "^2.0.5",
-    "eslint-loader": "^1.5.0",
-    "eslint-plugin-html": "^1.3.0",
-    "eslint-plugin-promise": "^3.4.0",
-    "eslint-plugin-standard": "^2.0.1",
-    "eslint-plugin-vue": "^1.0.0",
-    "eventsource-polyfill": "^0.9.6",
-    "express": "^4.13.3",
-    "extract-text-webpack-plugin": "^2.0.0",
-    "file-loader": "^0.9.0",
-    "friendly-errors-webpack-plugin": "^1.1.3",
-    "function-bind": "^1.0.2",
-    "html-webpack-plugin": "^2.28.0",
-    "http-proxy-middleware": "^0.17.2",
-    "inject-loader": "^2.0.1",
-    "json-loader": "^0.5.4",
-    "karma": "^1.4.1",
-    "karma-chrome-launcher": "^2.2.0",
-    "karma-coverage": "^1.1.1",
-    "karma-mocha": "^1.2.0",
-    "karma-sinon-chai": "^1.2.0",
-    "karma-sourcemap-loader": "^0.3.7",
-    "karma-spec-reporter": "0.0.26",
-    "karma-webpack": "^2.0.2",
-    "lolex": "^1.5.2",
-    "mocha": "^3.2.0",
-    "nightwatch": "^0.9.8",
-    "node-sass": "^4.5.3",
-    "normalize.css": "^5.0.0",
-    "opn": "^4.0.2",
-    "optimize-css-assets-webpack-plugin": "^1.3.0",
-    "ora": "^0.3.0",
-    "particles.js": "^2.0.0",
-    "sass-loader": "^6.0.3",
-    "script-loader": "^0.7.0",
-    "selenium-server": "2.53.1",
-    "semver": "^5.3.0",
-    "shelljs": "^0.7.4",
-    "sinon": "^2.1.0",
-    "sinon-chai": "^2.8.0",
-    "url-loader": "^0.5.8",
-    "vue": "^2.4.2",
-    "vue-loader": "^12.1.0",
-    "vue-markdown-loader": "^0.6.0",
-    "vue-router": "^2.2.0",
-    "vue-style-loader": "^2.0.0",
-    "vue-template-compiler": "^2.4.2",
-    "vuep": "^0.7.0",
-    "webpack": "^2.6.1",
-    "webpack-bundle-analyzer": "^2.2.1",
-    "webpack-dashboard": "^0.2.1",
-    "webpack-dev-middleware": "^1.10.0",
-    "webpack-hot-middleware": "^2.16.1",
-    "webpack-merge": "^2.6.1"
+    "@ant-design/icons-vue": "^6.0.1",
+    "@fe6/norm": "^3.2.0",
+    "@fe6/shared": "^0.9.1",
+    "@fe6/water-pro": "^3.0.29",
+    "@types/lodash-es": "^4.17.4",
+    "@types/node": "^15.12.5",
+    "@vitejs/plugin-vue": "^1.2.4",
+    "@vitejs/plugin-vue-jsx": "^1.1.6",
+    "@vue/compiler-sfc": "^3.1.2",
+    "axios": "^0.21.1",
+    "dotenv": "^10.0.0",
+    "husky": "^6.0.0",
+    "lodash-es": "^4.17.21",
+    "sass": "^1.35.1",
+    "typescript": "^4.3.4",
+    "vite": "^2.3.8",
+    "vite-plugin-compression": "^0.2.5",
+    "vite-plugin-html": "^2.0.7",
+    "vite-plugin-imagemin": "^0.3.2",
+    "vite-plugin-md": "^0.6.7",
+    "vite-plugin-style-import": "^1.0.1",
+    "vue": "^3.1.2",
+    "vue-router": "^4.0.10",
+    "vuex": "^4.0.2",
+    "yargs": "^17.0.1"
   },
-  "dependencies": {
-    "uppercamelcase": "^1.1.0"
+  "config": {
+    "commitizen": {
+      "path": "./node_modules/cz-customizable"
+    },
+    "cz-customizable": {
+      "config": "./node_modules/@fe6/norm/src/.cz-configrc.js"
+    }
+  },
+  "publishConfig": {
+    "access": "public"
   },
   "engines": {
-    "core-js": "^2.5.0",
-    "node": ">= 4.0.0",
-    "npm": ">= 3.0.0"
+    "node": ">=16.0.0"
+  },
+  "dependencies": {
+    "mitt": "^3.0.0",
+    "uppercamelcase": "^3.0.0"
   }
 }
