@@ -5,7 +5,7 @@ import registerMixin from '../mixins/register-component';
 import editorMixin from '../mixins/editor-component';
 import { lngLatTo } from '../utils/convert-helper';
 export default {
-  name: 'el-amap-polygon',
+  name: 'ElAmapPolygon',
   mixins: [registerMixin, editorMixin],
   props: [
     'vid',
@@ -33,7 +33,7 @@ export default {
           flag === false ? this.hide() : this.show();
         },
         zIndex(num) {
-          this.setOptions({zIndex: num});
+          this.setOptions({ zIndex: num });
         },
         editable(flag) {
           flag === true ? this.editor.open() : this.editor.close();
@@ -43,7 +43,7 @@ export default {
   },
   methods: {
     initComponent() {
-      let options = this.convertProps();
+      const options = this.convertProps();
       this.$amapComponent = new AMap.Polygon(options);
       this.$amapComponent.editor = new AMap.PolyEditor(this.$amap, this.$amapComponent);
     },
@@ -54,7 +54,7 @@ export default {
       return this.$amapComponent.getExtData();
     },
     $$contains(point) {
-      if (Array.isArray(point)) point = new AMap.LngLat(point[0], point[1]);
+      if (Array.isArray(point)) { point = new AMap.LngLat(point[0], point[1]); }
       return this.$amapComponent.getBounds().contains(point);
     }
   }
