@@ -5,7 +5,7 @@ import registerMixin from '../mixins/register-component';
 import editorMixin from '../mixins/editor-component';
 import { lngLatTo } from '../utils/convert-helper';
 export default {
-  name: 'el-amap-polyline',
+  name: 'ElAmapPolyline',
   mixins: [registerMixin, editorMixin],
   props: [
     'vid',
