@@ -11,13 +11,13 @@ export function assign(target, varArgs) {
       throw new TypeError('Cannot convert undefined or null to object');
     }
 
-    var to = Object(target);
+    const to = Object(target);
 
-    for (var index = 1; index < arguments.length; index++) {
-      var nextSource = arguments[index];
+    for (let index = 1; index < arguments.length; index++) {
+      const nextSource = arguments[index];
 
       if (nextSource != null) { // Skip over if undefined or null
-        for (var nextKey in nextSource) {
+        for (const nextKey in nextSource) {
           // Avoid bugs when hasOwnProperty is shadowed
           if (Object.prototype.hasOwnProperty.call(nextSource, nextKey)) {
             to[nextKey] = nextSource[nextKey];
@@ -26,7 +26,8 @@ export function assign(target, varArgs) {
       }
     }
     return to;
-  } else {
+  }
+  else {
     return Object.assign.apply(Object, arguments);
   }
-};
+}
